Add sorting of child nodes in howtotree.cpp, ignore binary. Move suffix tree tests and use new cpp order.
