+++ cpp/sorted.txt
@@ -1,0 +1,65 @@
+Suffix tree for: x
+1   |            | Root   | Root   | No SL  | [2]
+2   | x          | 0      | End    | No SL  | []
+Suffix tree for: xa
+1   |            | Root   | Root   | No SL  | [2, 3]
+2   | xa         | 0      | End    | No SL  | []
+3   | a          | 1      | End    | No SL  | []
+Suffix tree for: xac
+1   |            | Root   | Root   | No SL  | [2, 3, 4]
+2   | xac        | 0      | End    | No SL  | []
+3   | ac         | 1      | End    | No SL  | []
+4   | c          | 2      | End    | No SL  | []
+Suffix tree for: xacc
+1   |            | Root   | Root   | No SL  | [2, 3, 4]
+2   | xacc       | 0      | End    | No SL  | []
+3   | acc        | 1      | End    | No SL  | []
+4   | cc         | 2      | End    | No SL  | []
+Suffix tree for: xaccx
+1   |            | Root   | Root   | No SL  | [2, 3, 5]
+2   | xaccx      | 0      | End    | No SL  | []
+3   | accx       | 1      | End    | No SL  | []
+4   | cx         | 3      | End    | No SL  | []
+5   | c          | 2      | 3      | 1      | [4, 6]
+6   | x          | 4      | End    | No SL  | []
+Suffix tree for: xaccxa
+1   |            | Root   | Root   | No SL  | [2, 3, 5]
+2   | xaccxa     | 0      | End    | No SL  | []
+3   | accxa      | 1      | End    | No SL  | []
+4   | cxa        | 3      | End    | No SL  | []
+5   | c          | 2      | 3      | 1      | [4, 6]
+6   | xa         | 4      | End    | No SL  | []
+Suffix tree for: xaccxac
+1   |            | Root   | Root   | No SL  | [2, 3, 5]
+2   | xaccxac    | 0      | End    | No SL  | []
+3   | accxac     | 1      | End    | No SL  | []
+4   | cxac       | 3      | End    | No SL  | []
+5   | c          | 2      | 3      | 1      | [4, 6]
+6   | xac        | 4      | End    | No SL  | []
+Suffix tree for: xaccxaca
+1   |            | Root   | Root   | No SL  | [5, 7, 9]
+2   | cxaca      | 3      | End    | No SL  | []
+3   | cxaca      | 3      | End    | No SL  | []
+4   | cxaca      | 3      | End    | No SL  | []
+5   | c          | 2      | 3      | 1      | [4, 6, 11]
+6   | xaca       | 4      | End    | No SL  | []
+7   | xac        | 0      | 3      | 9      | [2, 8]
+8   | a          | 7      | End    | No SL  | []
+9   | ac         | 1      | 3      | 5      | [3, 10]
+10  | a          | 7      | End    | No SL  | []
+11  | a          | 7      | End    | No SL  | []
+Suffix tree for: xaccxaca$
+1   |            | Root   | Root   | No SL  | [5, 7, 12, 14]
+2   | cxaca$     | 3      | End    | No SL  | []
+3   | cxaca$     | 3      | End    | No SL  | []
+4   | cxaca$     | 3      | End    | No SL  | []
+5   | c          | 2      | 3      | 1      | [4, 6, 11]
+6   | xaca$      | 4      | End    | No SL  | []
+7   | xac        | 0      | 3      | 9      | [2, 8]
+8   | a$         | 7      | End    | No SL  | []
+9   | c          | 2      | 3      | 5      | [3, 10]
+10  | a$         | 7      | End    | No SL  | []
+11  | a$         | 7      | End    | No SL  | []
+12  | a          | 1      | 2      | 1      | [9, 13]
+13  | $          | 8      | End    | No SL  | []
+14  | $          | 8      | End    | No SL  | []
