+++ src/data_structures/suffix_trees.rs
@@ -310,3 +310,39 @@ impl fmt::Display for Node {
         write!(f, "{:<6} | {:<6} | {:<6} | {:?}", start, end, sl, children)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn create_suffix_trees_1() {
+        let st = SuffixTree::new("x");
+        let expected_st =
+"Suffix tree for: x
+1   |            | Root   | Root   | No SL  | [2]
+2   | x          | 0      | End    | No SL  | []
+";
+        assert_eq!(expected_st, format!("{}", st));
+    }
+
+    #[test]
+    fn create_suffix_trees_2() {
+        let st = SuffixTree::new("xaccxaca");
+        let expected_st =
+    "Suffix tree for: xaccxaca
+1   |            | Root   | Root   | No SL  | [5, 7, 9]
+2   | cxaca      | 3      | End    | No SL  | []
+3   | cxaca      | 3      | End    | No SL  | []
+4   | cxaca      | 3      | End    | No SL  | []
+5   | c          | 2      | 3      | 1      | [4, 6, 11]
+6   | xaca       | 4      | End    | No SL  | []
+7   | xac        | 0      | 3      | 9      | [2, 8]
+8   | a          | 7      | End    | No SL  | []
+9   | ac         | 1      | 3      | 5      | [3, 10]
+10  | a          | 7      | End    | No SL  | []
+11  | a          | 7      | End    | No SL  | []
+";
+        assert_eq!(expected_st, format!("{}", st));
+    }
+}
