+++ tests/tests.rs
@@ -1,5 +1,4 @@
 // Data structures tests
 mod data_structures {
     mod edit_distance_tests;
-    mod suffix_tree_tests;
 }
